0215 - greengram5 JPA2

+++ src/main/java/com/green/greengram4/user/model/UserFollowDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class UserFollowDto {
-    private int fromIuser;
-    private int toIuser;
+    private long fromIuser;
+    private long toIuser;
 }
